Reduced download time if src.rpm already present.

+++ autotest/extractpot.py
@@ -15,7 +15,7 @@ class ExtractPot:
        Koji and unpacking the rpm and extracting the .pot file.
        Then translation statistics are generated.
     """
-    
+
     def __init__(self, program_name, locale):
         self.program_name = program_name
         self.locale = locale
@@ -30,10 +30,10 @@ class ExtractPot:
 
         self.srpm_dir = self.home+"/.autotest/srpm"
         self.srpm_path = self.srpm_dir+"/"+self.program_name+".src.rpm"
-        
+
         if not os.path.exists(self.home+"/.autotest/pot_files"):
             os.makedirs(self.home+"/.autotest/pot_files")
-    
+
         if not os.path.exists(self.home+"/.autotest/srpm"):
             os.makedirs(self.home+"/.autotest/srpm")
 
@@ -43,7 +43,7 @@ class ExtractPot:
            Unpack the downloaded RPM and generate the .pot file.
            .pot file is then moved to ~/.autotest/pot_files.
         """
-        
+
         top_url = "https://kojipkgs.fedoraproject.org/packages"
         pkg_path = subprocess.check_output(['rpm -qa '+self.program_name+' --queryformat %{NAME}/%{VERSION}/%{RELEASE}'], shell=True)
         pkg_name = subprocess.check_output(['rpm -qa '+self.program_name+' --queryformat %{NAME}-%{VERSION}-%{RELEASE}'], shell=True)
@@ -51,8 +51,11 @@ class ExtractPot:
         pkg_url = top_url+"/"+pkg_path+"/src/"+pkg_name+".src.rpm"
         log.info("Download URL: %s", pkg_url)
 
-        urllib.urlretrieve(pkg_url, self.srpm_path)
-        log.info("src.rpm downloaded.")
+        if not os.path.isfile(self.srpm_path):
+            urllib.urlretrieve(pkg_url, self.srpm_path)
+            log.info("src.rpm downloaded.")
+        else:
+            log.info("src.rpm already present, not downloading.")
 
         os.chdir(self.srpm_dir)
 
@@ -65,9 +68,9 @@ class ExtractPot:
 
         command = "tar -xf "+tar_name+".tar.xz"
         command = shlex.split(command)
-        subprocess.call(command) 
+        subprocess.call(command)
         log.info("src.rpm unpacked in %s", self.srpm_dir)
-        
+
         os.chdir(self.srpm_dir+"/"+tar_name+"/po")
 
         command = "intltool-update -p"
@@ -90,7 +93,7 @@ class ExtractPot:
         os.chdir(path)
         po_file = self.locale+"."+self.program_name+".po"
         pot_name = self.program_name+".pot"
-        
+
         command = "intltool-update --dist --gettext-package="+self.program_name+" "+self.locale+"."+self.program_name
         command = shlex.split(command)
         subprocess.call(command)
@@ -100,9 +103,9 @@ class ExtractPot:
         command = shlex.split(command)
         stats = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         out, err = stats.communicate()
-        
+
         os.chdir(os.path.abspath(os.path.dirname(__file__)))
-        
+
         return err
